+++ src/js/core/height-viewport.js
@@ -65,7 +65,10 @@ export default {
 
                 if (this.offsetTop) {
                     if (isScrollingElement) {
-                        const top = offsetPosition(this.$el)[0] - offsetPosition(scrollElement)[0];
+                        const offsetTopEl =
+                            this.offsetTop === true ? this.$el : query(this.offsetTop, this.$el);
+                        const top =
+                            offsetPosition(offsetTopEl)[0] - offsetPosition(scrollElement)[0];
                         minHeight += top > 0 && top < viewportHeight / 2 ? ` - ${top}px` : '';
                     } else {
                         minHeight += ` - ${css(scrollElement, 'paddingTop')}`;